requirements.txt
mcp>=1.0.0
neo4j>=5.15.0
rich>=13.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
asyncio-mqtt>=0.11.1